examples/simple-starlette-demo/requirements.txt
gino-starlette==0.1.0a0
starlette==0.13.1
uvicorn==0.11.2